Lazypanda/backend/requirements.txt
Flask
flask-pymongo
flask-cors
scrapy
werkzeug
PyJWT
